division exception added

+++ src/Calc.java
@@ -31,6 +31,8 @@ public class Calc {
         return num1 * num2;
     }
     public double division(){
-        return num1 / num2;
+        return (double) num1/num2;
     }
+
+
 }
